requirements.txt
slack-bolt>=1.18.0
spotipy>=2.23.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
